--- src/main/java/com/github/siboxd/fatturapa/model/EsigibilitaIVAType.java
@@ -1,14 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Namespace;
-import org.simpleframework.xml.Root;
-
-@Root(name = "EsigibilitaIVAType")
-@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
-public enum EsigibilitaIVAType {
-
-    D,
-    I,
-    S
-
-}
